Move server to MSVC2008

+++ src_main/dlls/TemplateEntities.cpp
@@ -274,7 +274,7 @@ void Templates_StartUniqueInstance( void )
 	g_iCurrentTemplateInstance++;
 
 	// Make sure there's enough room to fit it into the string
-	int iMax = pow(10, (strlen(ENTITYIO_FIXUP_STRING)-1));	// -1 for the &
+	int iMax = pow((float)10, (float)(strlen(ENTITYIO_FIXUP_STRING)-1));	// -1 for the &
 	if ( g_iCurrentTemplateInstance >= iMax )
 	{
 		// We won't hit this.
